// src/types/types.ts
export interface LLMConfig {
    model: string;
    base_url?: string;
    api_key: string;
    temperature?: number;
    api_type: string;
  }

export interface EmbeddingsConfig {
    model: string;
    base_url?: string;
    api_key?: string;
    huggingface_model?: string;
    embedding_type: string;
}
  
  export interface SQLConfig {
    url: string;
    username: string;
    password: string;
    db_name: string;
  }
  
  export interface S3Config {
    bucket_name: string;
    region_name: string;
    aws_access_key: string;
    aws_secret_key: string;
  }
  
  export interface RAGConfig {
    llm: string;
    embeddings_model: string;
    collection: string;
    system_prompt?: string;
    contextualization_prompt?: string;
    temperature?: number;
    advancedLLMConfig?: LLMConfig;
    sql_config?: SQLConfig;
    s3_config?: S3Config;
    advancedEmbeddingsConfig?: EmbeddingsConfig;
  }
  
  export interface Agent {
    id: string;
    user_id: string;
    config: RAGConfig;
    created_at: string;
  }
// ...
  export interface EvalAgent {
    id: string;
    config: {
      collection: string;
      llm: string;
    };
    status?: 'active' | 'inactive'| string;
    lastActive?: string;
  }
  
  export type EvaluationStatus = 'processing' | 'completed' | 'failed';
  export type FilterStatus = 'all' | 'active' | 'completed' | 'failed';
  
  export interface EvaluationJob {
    _id: string;
    agent_id: string;
    status: EvaluationStatus;
    progress: number;
    total_questions: number;
    processed_questions: number;
    created_at: string;
    completion_time?: string;
    error?: string;
  }
  
  export interface EvaluationResult {
    question: string;
    original_answer: string;
    generated_answer: string;
    similarity_score: number;
  }
  
  export interface AggregateMetrics {
    mean_similarity: number;
    median_similarity: number;
    min_similarity: number;
    max_similarity: number;
    std_similarity: number;
  }
  
  export interface Evaluation {
    agent_id: string;
    job_id: string;
    timestamp: string;
    status: EvaluationStatus;
    results: EvaluationResult[];
    aggregate_metrics: AggregateMetrics;
    error?: string;
  }